SA-LSA

+++ MFEA_lib/operators/Selection.py
@@ -19,7 +19,7 @@ class ElitismSelection(AbstractSelection):
         idx_selected_inds = np.empty((0,), dtype= int )
         
         for i in range (len(nb_inds_tasks)):
-            N_i = nb_inds_tasks[i]
+            N_i = np.int(nb_inds_tasks[i])
 
             idx_inds_i = np.where(skill_factor_arr == i)[0]
             sorted_idx = idx_inds_i[np.argsort(-pop_fitness[idx_inds_i])]
